Removed an ugly cast in socket.c Fixed Makefile (-Wno-unused-parameter was added twice with gcc4)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@10205 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/socket.h
@@ -121,7 +121,7 @@ void socket_final(void);
 
 extern void flush_fifo(int fd);
 extern void flush_fifos(void);
-extern void set_nonblocking(int fd, int yes);
+extern void set_nonblocking(int fd, unsigned long yes);
 
 void set_defaultparse(ParseFunc defaultparse);
